--- rmb-sdk-go/common/README.md
@@ -1,8 +1,0 @@
-# Common
-
-Contains all the shared logic between the different RMB clients.
-
-## Types generation
-```bash
-protoc -I. --go_out=types types.proto
-```
