+++ rmb-sdk-go/connection.go
@@ -1,4 +1,4 @@
-package common
+package rmb
 
 import (
 	"context"
