+++ rmb-sdk-go/examples/client_async/main.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	substrate "github.com/threefoldtech/tfchain/clients/tfchain-client-go"
+	"github.com/threefoldtech/tfgrid-sdk-go/rmb-sdk-go"
 	"github.com/threefoldtech/tfgrid-sdk-go/rmb-sdk-go/async"
-	"github.com/threefoldtech/tfgrid-sdk-go/rmb-sdk-go/common"
 )
 
 // Change result to the expected response
@@ -33,7 +33,7 @@ func app() error {
 	}
 
 	defer sub.Close()
-	client, err := async.NewAsyncClient(context.Background(), listener, common.KeyTypeSr25519, mnemonics, "wss://relay.dev.grid.tf", "test-client", sub, true)
+	client, err := async.NewAsyncClient(context.Background(), listener, rmb.KeyTypeSr25519, mnemonics, "wss://relay.dev.grid.tf", "test-client", sub, true)
 	if err != nil {
 		return fmt.Errorf("failed to create direct client: %w", err)
 	}
