+++ rmb-sdk-go/examples/client_direct/main.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	substrate "github.com/threefoldtech/tfchain/clients/tfchain-client-go"
-	"github.com/threefoldtech/tfgrid-sdk-go/rmb-sdk-go/common"
+	"github.com/threefoldtech/tfgrid-sdk-go/rmb-sdk-go"
 	"github.com/threefoldtech/tfgrid-sdk-go/rmb-sdk-go/direct"
 )
 
@@ -20,7 +20,7 @@ func app() error {
 	}
 
 	defer sub.Close()
-	client, err := direct.NewClient(context.Background(), common.KeyTypeSr25519, mnemonics, "wss://relay.dev.grid.tf", "test-client", sub, true)
+	client, err := direct.NewClient(context.Background(), rmb.KeyTypeSr25519, mnemonics, "wss://relay.dev.grid.tf", "test-client", sub, true)
 	if err != nil {
 		return fmt.Errorf("failed to create direct client: %w", err)
 	}
