+++ rmb-sdk-go/twindb.go
@@ -1,4 +1,4 @@
-package common
+package rmb
 
 import (
 	"fmt"
